+++ ChangeLog.md
@@ -13,6 +13,8 @@
 
   * Add per-assessment PrairieSchedule exam linking (Matt West).
 
+  * Add "Report cheating" page in Exam mode (Matt West).
+
   * Fix migration 111 to allow re-running (Matt West).
 
 * 2.12.0 - 2018-05-19